+++ rosie/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this plugin will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this plugin adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2025-12-12
+
+### Added
+- Prefabs can now be edited using the same editing instructions as the scene.
+- The plugin will block accessing certain files until the correct skill has been used.
+
+### Changed
+- Properties of Game Objects and Components can now be set at the time they are created.
+- Updated some instructions from Drew's feedback.
+
 ## [0.3.1] - 2025-12-08
 
 ### Added